Remove --no-cache option when building container images
This isn't necessary, as the commands in our Dockerfile aren't the
kind that would erroneously *keep* the cache valid (e.g., `RUN apt-get
-y update`).

Furthermore, our Dockerfile is generated automatically via template
runtime based on the packages being installed, so the text of those
commands, as well as the contents of the files used in `ADD` commands
will be changing anyway.

This allows us to start getting some reuse out of our container
creation workflow!

Signed-off-by: Christopher Maier <cmaier@chef.io>